+++ src/session/index.ts
@@ -3,3 +3,6 @@ export { Inserter } from "./Inserter";
 export { Query } from "./Query";
 export { Paginator } from "./Paginator";
 export { Session } from "./Session";
+export type { AgentSession } from "./AgentSession";
+export { ListSession } from "./ListSession";
+export type { ListRow, ListCandidate } from "./ListSession";